chore: move section with logos to main page
