+++ src/pages/index.jsx
@@ -10,10 +10,73 @@ import Reviews from 'components/pages/home/reviews';
 import Layout from 'components/shared/layout';
 import CtaWithForm from 'components/shared/reusable-sections/cta-with-form/cta-with-form';
 import Inbox from 'components/shared/reusable-sections/inbox';
+import SectionWithLogos from 'components/shared/reusable-sections/section-with-logos';
 import SEO from 'components/shared/seo';
 import Separator from 'components/shared/separator';
-
+import allstarLogo from 'images/reusable-sections/section-with-logos/allstar.svg';
+import axiosHqLogo from 'images/reusable-sections/section-with-logos/axios-hq.svg';
+import baskLogo from 'images/reusable-sections/section-with-logos/bask.svg';
+import capgeminiLogo from 'images/reusable-sections/section-with-logos/capgemini.svg';
+import mongoDbLogo from 'images/reusable-sections/section-with-logos/mongodb.svg';
+import mothershipLogo from 'images/reusable-sections/section-with-logos/mothership.svg';
+import rocheLogo from 'images/reusable-sections/section-with-logos/roche.svg';
+import runnLogo from 'images/reusable-sections/section-with-logos/runn.svg';
+import saladLogo from 'images/reusable-sections/section-with-logos/salad.svg';
+import siemensLogo from 'images/reusable-sections/section-with-logos/siemens.svg';
+import teocoLogo from 'images/reusable-sections/section-with-logos/teoco.svg';
+import unityLogo from 'images/reusable-sections/section-with-logos/unity.svg';
 // import LINKS from 'constants/links';
+
+const SECTION_WITH_LOGOS_2 = [
+  {
+    title: 'Salad',
+    src: saladLogo,
+  },
+  {
+    title: 'MongoDB',
+    src: mongoDbLogo,
+  },
+  {
+    title: 'Mothership',
+    src: mothershipLogo,
+  },
+  {
+    title: 'Capgemini',
+    src: capgeminiLogo,
+  },
+  {
+    title: 'Bask',
+    src: baskLogo,
+  },
+  {
+    title: 'Roche',
+    src: rocheLogo,
+  },
+  {
+    title: 'Unity',
+    src: unityLogo,
+  },
+  {
+    title: 'Siemens',
+    src: siemensLogo,
+  },
+  {
+    title: 'Teoco',
+    src: teocoLogo,
+  },
+  {
+    title: 'Axio HQ',
+    src: axiosHqLogo,
+  },
+  {
+    title: 'Runn',
+    src: runnLogo,
+  },
+  {
+    title: 'Allstar',
+    src: allstarLogo,
+  },
+];
 
 const INBOX_CATEGORIES = ["What's New", 'Alerts', 'Account'];
 
@@ -183,6 +246,12 @@ const INBOX_MESSAGES = [
 const HomePage = () => (
   <Layout mainClassName="reusable-components overflow-hidden pt-16 bg-[#05050B]">
     <Hero />
+    <SectionWithLogos
+      containerSize="lg"
+      title="Notifications That Brands Count On"
+      description="Ensuring seamless notifications from business to users, with zero hassle."
+      logos={SECTION_WITH_LOGOS_2}
+    />
     <Integration />
     <Inbox
       title="The most customizable <Inbox&nbsp;/>"
